requirements.txt
selenium==3.141.0
PyYAML==3.13
